Extract provision

+++ provision.sh
@@ -1,10 +1,16 @@
 #!/bin/bash
 
 # fail immediately on error
-set -e
+set -e -x
+
+# echo "$0 $*" > ~/provision.log
+
+fail() {
+  echo "$*" >&2
+  exit 1
+}
 
 # Variables passed in from terraform, see openstack-cf-install.tf, the "remote-exec" provisioner
-
 OS_USERNAME=${1}
 OS_API_KEY=${2}
 OS_TENANT=${3}
@@ -22,6 +28,16 @@ INSTALL_DOCKER=${13}
 
 boshDirectorHost="${IPMASK}.2.4"
 cfReleaseVersion="207"
+
+cd $HOME
+(("$?" == "0")) ||
+  fail "Could not find HOME folder, terminating install."
+
+# Generate the key that will be used to ssh between the bastion and the
+# microbosh machine
+if [[ ! -f ~/.ssh/id_rsa ]]; then
+  ssh-keygen -t rsa -N "" -f ~/.ssh/id_rsa
+fi
 
 # Prepare the jumpbox to be able to install ruby and git-based bosh and cf repos
 
@@ -50,39 +66,31 @@ esac
 
 cd $HOME
 
-# Generate the key that will be used to ssh between the bastion and the
-# microbosh machine
-ssh-keygen -t rsa -N "" -f ~/.ssh/id_rsa
 
 
 # Install RVM
-# gpg --keyserver hkp://keys.gnupg.net --recv-keys 409B6B1796C275462A1703113804BB82D39DC0E3
-# curl -sSL https://get.rvm.io | bash -s stable
 
-git clone git://github.com/rvm/rvm
-cd rvm
-./install
+if [[ ! -d "$HOME/rvm" ]]; then
+  git clone git://github.com/rvm/rvm
+fi
+
+if [[ ! -d "$HOME/.rvm" ]]; then
+  cd rvm
+  ./install
+fi
+
 cd $HOME
 
-~/.rvm/bin/rvm install ruby-2.1
-~/.rvm/bin/rvm alias create default 2.1
+if [[ ! "$(ls -A $HOME/.rvm/environments)" ]]; then
+  ~/.rvm/bin/rvm install ruby-2.1
+fi
+
+if [[ ! -d "$HOME/.rvm/environments/default" ]]; then
+  ~/.rvm/bin/rvm alias create default 2.1
+fi
+
 source ~/.rvm/environments/default
 source ~/.rvm/scripts/rvm
-
-
-# This volume is created using terraform in aws-bosh.tf
-#sudo /sbin/mkfs.ext4 /dev/xvdc
-#sudo /sbin/e2label /dev/xvdc workspace
-#echo 'LABEL=workspace /home/ubuntu/workspace ext4 defaults,discard 0 0' | sudo tee -a /etc/fstab
-#mkdir -p /home/ubuntu/workspace
-#sudo mount -a
-#sudo chown -R ubuntu:ubuntu /home/ubuntu/workspace
-
-# As long as we have a large volume to work with, we'll move /tmp over there
-# You can always use a bigger /tmp
-#sudo rsync -avq /tmp/ /home/ubuntu/workspace/tmp/
-#sudo rm -fR /tmp
-#sudo ln -s /home/ubuntu/workspace/tmp /tmp
 
 # Install BOSH CLI, bosh-bootstrap, spiff and other helpful plugins/tools
 gem install fog-aws -v 0.1.1 --no-ri --no-rdoc --quiet
@@ -113,20 +121,28 @@ address:
   ip: ${boshDirectorHost}
 EOF
 
-bosh bootstrap deploy
+if [[ ! -d "$HOME/workspace/deployments/microbosh/deployments" ]]; then
+  bosh bootstrap deploy
+fi
 
 # We've hardcoded the IP of the microbosh machine, because convenience
 bosh -n target https://${boshDirectorHost}:25555
 bosh login admin admin
+
+if [[ ! "$?" == 0 ]]; then
+  #wipe the ~/workspace/deployments/microbosh folder contents and try again
+  echo "Retry deploying the micro bosh..."
+fi
 popd
 
 # There is a specific branch of cf-boshworkspace that we use for terraform. This
 # may change in the future if we come up with a better way to handle maintaining
 # configs in a git repo
-git clone --branch  ${CF_BOSHWORKSPACE_VERSION} http://github.com/cloudfoundry-community/cf-boshworkspace
+if [[ ! -d "$HOME/workspace/deployments/cf-boshworkspace" ]]; then
+  git clone --branch  ${CF_BOSHWORKSPACE_VERSION} http://github.com/cloudfoundry-community/cf-boshworkspace
+fi
 pushd cf-boshworkspace
 mkdir -p ssh
-
 gem install bundler
 bundle install
 
@@ -139,10 +155,12 @@ if [ $CF_DOMAIN == "XIP" ]; then
   CF_DOMAIN="${CF_IP}.xip.io"
 fi
 
-curl -sOL https://github.com/cloudfoundry-incubator/spiff/releases/download/v1.0.3/spiff_linux_amd64.zip
-unzip spiff_linux_amd64.zip
-sudo mv ./spiff /usr/local/bin/spiff
-rm spiff_linux_amd64.zip
+if [[ ! -f "/usr/local/bin/spiff" ]]; then
+  curl -sOL https://github.com/cloudfoundry-incubator/spiff/releases/download/v1.0.3/spiff_linux_amd64.zip
+  unzip spiff_linux_amd64.zip
+  sudo mv ./spiff /usr/local/bin/spiff
+  rm spiff_linux_amd64.zip
+fi
 
 # This is some hackwork to get the configs right. Could be changed in the future
 /bin/sed -i \
@@ -159,30 +177,46 @@ rm spiff_linux_amd64.zip
 
 
 # Upload the bosh release, set the deployment, and execute
-bosh upload release https://bosh.io/d/github.com/cloudfoundry/cf-release?v=${cfReleaseVersion}
-bosh deployment cf-openstack-${CF_SIZE}
-bosh prepare deployment
+deployedVersion=$(bosh releases | grep " ${cfReleaseVersion}" | awk '{print $4}')
+deployedVersion="${deployedVersion//[^[:alnum:]]/}"
+if [[ ! "$deployedVersion" == "${cfReleaseVersion}" ]]; then
+  bosh upload release https://bosh.io/d/github.com/cloudfoundry/cf-release?v=${cfReleaseVersion}
+  bosh deployment cf-openstack-${CF_SIZE}
+  bosh prepare deployment || bosh prepare deployment  #Seems to always fail on the first run...
+else
+  bosh deployment cf-openstack-${CF_SIZE}
+fi
+
+# Work around until bosh-workspace can handle submodules
+if [[ "cf-openstack-${CF_SIZE}" == "cf-openstack-large" ]]; then
+  pushd .releases/cf
+  ./update
+  popd
+fi
 
 # We locally commit the changes to the repo, so that errant git checkouts don't
 # cause havok
 #git commit -am 'commit of the local deployment configs'
 
-# Speaking of hack-work, bosh deploy often fails the first or even second time, due to packet bats
-# We run it three times (it's idempotent) so that you don't have to
+# Keep trying until there is a successful BOSH deploy.
 for i in {0..2}
 do bosh -n deploy
 done
 
-
 echo "Install Traveling CF"
-curl -s https://raw.githubusercontent.com/cloudfoundry-community/traveling-cf-admin/master/scripts/installer | bash
-echo 'export PATH=$PATH:$HOME/bin/traveling-cf-admin' >> ~/.bashrc
+if [[ "$(cat $HOME/.bashrc | grep 'export PATH=$PATH:$HOME/bin/traveling-cf-admin')" == "" ]]; then
+  curl -s https://raw.githubusercontent.com/cloudfoundry-community/traveling-cf-admin/master/scripts/installer | bash
+  echo 'export PATH=$PATH:$HOME/bin/traveling-cf-admin' >> $HOME/.bashrc
+  source $HOME/.bashrc
+fi
 
 # Now deploy docker services if requested
 if [[ $INSTALL_DOCKER == "true" ]]; then
 
   cd ~/workspace/deployments
-  git clone https://github.com/cloudfoundry-community/docker-services-boshworkspace.git
+  if [[ ! -d "$HOME/workspace/deployments/docker-services-boshworkspace" ]]; then
+    git clone https://github.com/cloudfoundry-community/docker-services-boshworkspace.git
+  fi
 
   echo "Update the docker-aws-vpc.yml with cf-boshworkspace parameters"
   /home/ubuntu/workspace/deployments/docker-services-boshworkspace/shell/populate-docker-openstack
@@ -200,6 +234,9 @@ if [[ $INSTALL_DOCKER == "true" ]]; then
   done
 
 fi
+
+echo "Provision script completed..."
+exit 0
 
 # FIXME: enable this again when smoke_tests work
 # bosh run errand smoke_tests
